+++ tests/AlternateProviderTest.php
@@ -1,0 +1,67 @@
+<?php
+
+/**
+ * This file is part of Laravel TestBench by Graham Campbell.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace GrahamCampbell\Tests\TestBench;
+
+use GrahamCampbell\TestBench\AbstractLaravelTestCase;
+use GrahamCampbell\TestBench\Traits\ServiceProviderTestCaseTrait;
+
+/**
+ * This is the alternate provider test class.
+ *
+ * @package    Laravel-TestBench
+ * @author     Graham Campbell
+ * @copyright  Copyright 2013-2014 Graham Campbell
+ * @license    https://github.com/GrahamCampbell/Laravel-TestBench/blob/master/LICENSE.md
+ * @link       https://github.com/GrahamCampbell/Laravel-TestBench
+ */
+class AlternateProviderTest extends AbstractLaravelTestCase
+{
+    /**
+     * Get the application base path.
+     *
+     * @return string
+     */
+    protected function getBasePath()
+    {
+        return __DIR__.'/../src';
+    }
+
+    /**
+     * Specify if routing filters are enabled.
+     *
+     * @return bool
+     */
+    protected function enableFilters()
+    {
+        return true;
+    }
+
+    /**
+     * Get the required service providers.
+     *
+     * @return array
+     */
+    protected function getRequiredServiceProviders()
+    {
+        return array();
+    }
+
+    public function testFilters()
+    {
+        $this->assertTrue($this->isFiltering(), 'Expected filtering to be enabled.');
+    }
+}