+++ tests/FacadeTest.php
@@ -1,0 +1,64 @@
+<?php
+
+/**
+ * This file is part of Laravel TestBench by Graham Campbell.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace GrahamCampbell\Tests\TestBench;
+
+use GrahamCampbell\Tests\TestBench\AbstractTestCase;
+use GrahamCampbell\TestBench\Traits\FacadeTestCaseTrait;
+
+/**
+ * This is the facade test class.
+ *
+ * @package    Laravel-TestBench
+ * @author     Graham Campbell
+ * @copyright  Copyright 2013-2014 Graham Campbell
+ * @license    https://github.com/GrahamCampbell/Laravel-TestBench/blob/master/LICENSE.md
+ * @link       https://github.com/GrahamCampbell/Laravel-TestBench
+ */
+class FacadeTest extends AbstractTestCase
+{
+    use FacadeTestCaseTrait;
+
+    /**
+     * Get the facade accessor.
+     *
+     * @return string
+     */
+    protected function getFacadeAccessor()
+    {
+        return 'testbench.foostub';
+    }
+
+    /**
+     * Get the facade class.
+     *
+     * @return string
+     */
+    protected function getFacadeClass()
+    {
+        return 'GrahamCampbell\Tests\TestBench\FooFacade';
+    }
+
+    /**
+     * Get the facade route.
+     *
+     * @return string
+     */
+    protected function getFacadeRoot()
+    {
+        return 'GrahamCampbell\Tests\TestBench\FooStub';
+    }
+}